requirements.txt
alembic==1.18.3
anyio==4.12.1
certifi==2026.1.4
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-telegram-bot==22.6
respx==0.22.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
xmltodict==1.0.2
